@@ -19,10 +19,7 @@
     "release": "pkg --targets node16-linux-x64 --output dist/bin/webgrep .",
     "serve": "node dist/server",
     "dev": "concurrently --kill-others --prefix name --prefix-colors cyan.bold,magenta.bold --names HTTP,JS 'npm run start:server:watch' 'npm run build:client:watch'",
-    "lint": "eslint .",
-    "test": "jest",
-    "test:watch": "npm run test -- --watch",
-    "cover": "npm run test -- --coverage"
+    "lint": "eslint ."
   },
   "repository": {
     "type": "git",
@@ -31,8 +28,7 @@
   "author": "Kevin Lin <developer@kevinlin.info>",
   "license": "MIT",
   "pre-commit": [
-    "lint",
-    "test"
+    "lint"
   ],
   "pkg": {
     "assets": [
@@ -44,7 +40,7 @@
   "dependencies": {
     "@grpc/grpc-js": "^1.5.9",
     "@linkiwi/fonts": "https://lib.kevinlin.info/fonts/9d03f84000c20d139ad2a10a08c477fb0c38f62a/release.tgz",
-    "@linkiwi/hoc": "https://lib.kevinlin.info/hoc/5a44275bdf250d97fa5179928a1a93448c843118/release.tgz",
+    "@linkiwi/hoc": "https://lib.kevinlin.info/hoc/f1230dcc7bf0d2dd827bcea2d798428c4c950e3c/release.tgz",
     "@sentry/browser": "^6.18.2",
     "@sentry/node": "^6.18.2",
     "async": "^3.2.3",
@@ -59,8 +55,8 @@
     "livegrep": "https://lib.kevinlin.info/schemas/external/livegrep/559bc5cc628461667b88fea5a27c2b5d94076133/release.tgz",
     "node-statsd": "^0.1.1",
     "prop-types": "^15.8.1",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-elemental": "https://lib.kevinlin.info/react-elemental/1ad5838ec41b5778c4430c7e760e0e7eadc9c9bf/release.tgz",
     "react-favicon": "^0.0.26",
     "react-helmet": "^6.1.0",
@@ -74,7 +70,7 @@
     "redux": "^4.0.5",
     "redux-thunk": "^2.4.1",
     "snappyjs": "^0.6.0",
-    "supercharged": "https://lib.kevinlin.info/supercharged/50f69713ca1fb03d4b5397be0761a35bc6ebb58a/release.tgz",
+    "supercharged": "https://lib.kevinlin.info/supercharged/5d491a1de91b8b72ddb0eb29fe4c27318ccdc5b9/release.tgz",
     "winston": "^3.6.0",
     "yargs": "^17.4.0"
   },
@@ -88,8 +84,7 @@
     "@babel/preset-env": "^7.16.11",
     "@babel/preset-react": "^7.16.7",
     "@babel/register": "^7.17.7",
-    "@linkiwi/eslint-config": "https://lib.kevinlin.info/eslint-config/cbf6e408cd5706df66e770d2ca7960e20283b6a9/release.tgz",
-    "babel-jest": "^27.5.1",
+    "@linkiwi/eslint-config": "https://lib.kevinlin.info/eslint-config/688b3497ac64d2d3f5758febe71f6d8468d06c29/release.tgz",
     "babel-loader": "^8.1.0",
     "babel-plugin-module-resolver": "^4.0.0",
     "babel-plugin-react-elemental": "^1.2.3",
@@ -97,21 +92,14 @@
     "babel-plugin-transform-react-remove-prop-types": "^0.4.24",
     "concurrently": "^7.0.0",
     "copy-webpack-plugin": "^10.2.4",
-    "enzyme": "^3.11.0",
-    "enzyme-adapter-react-16": "^1.15.3",
     "eslint": "^8.11.0",
     "eslint-import-resolver-babel-module": "^5.1.2",
     "eslint-plugin-import": "^2.25.4",
     "eslint-plugin-jest": "^25.7.0",
     "eslint-plugin-react": "^7.29.4",
-    "identity-obj-proxy": "^3.0.0",
-    "jest": "^27.5.1",
-    "jest-environment-jsdom": "^27.5.1",
     "nodemon": "^2.0.15",
     "pkg": "^5.5.2",
     "pre-commit": "^1.2.2",
-    "raf": "^3.4.1",
-    "react-test-renderer": "^16.14.0",
     "terser-webpack-plugin": "^5.3.1",
     "url-loader": "^4.1.0",
     "webpack": "^5.70.0",
